Complete the result page, fix the filters, and added details to individual pages
